+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/_Template.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using Xamarin.Forms.CustomAttributes;
+﻿using Xamarin.Forms.CustomAttributes;
 using Xamarin.Forms.Internals;
 
 #if UITEST
@@ -32,54 +31,5 @@ namespace Xamarin.Forms.Controls.Issues
 			RunningApp.Screenshot ("I see the Label");
 		}
 #endif
-	}
-
-	[Preserve(AllMembers = true)]
-	[Issue(IssueTracker.Github, 1396, 
-		"Label TextAlignment is not kept when resuming application", 
-		PlatformAffected.Android)]
-	public class Issue1396Vertical : TestContentPage
-	{
-		Label _label;
-
-		protected override void Init()
-		{
-			var instructions = new Label
-			{
-				Text = "Tap the 'Change Text' button. Tap the Overview button. Resume the application. If the label" 
-						+ " text is no longer centered, the test has failed."
-			};
-
-			var button = new Button { Text = "Change Text" };
-			button.Clicked += (sender, args) =>
-			{
-				// TODO hartez 2018/03/29 09:34:21 We need to test this if it wraps 	
-
-				_label.Text = DateTime.Now.ToLongDateString() + " " + DateTime.Now.ToLongTimeString();
-			};
-
-			_label = new Label
-			{
-				HeightRequest = 400,
-				BackgroundColor = Color.Gold,
-				Text = "I should be centered in the gold area",
-				VerticalTextAlignment = TextAlignment.Center,
-				HorizontalTextAlignment = TextAlignment.Center
-			};
-
-			var layout = new StackLayout 
-			{
-				Children = { instructions, button, _label }
-			};
-
-			var content = new ContentPage 
-			{
-				Content = layout 
-			};
-
-			Content = new Label { Text = "Shouldn't see this" };
-
-			Appearing += (sender, args) => Application.Current.MainPage = content;
-		}
 	}
 }
